feat: add cliffy
