+++ deps/cliffy.ts
@@ -1,0 +1,1 @@
+export * from "https://deno.land/x/cliffy@v0.25.7/command/mod.ts"
